+++ consts/constants.py
@@ -1,9 +1,19 @@
 """
 consts/constants.py contains constants used throughout the bot.
 
+CONFIG_ID_MAP: A dictionary containing the valid ids and keys that can be selected to be incremented or in rare cases decremented in the config.ini file.
+
+CONFIG_ABSOLUTE_PATH: The absolute path to the config.ini file.
+
 GENERAL_CHANNEL, GOVERNANCE_CHANNEL, GOVERNANCE_BUDGET_CHANNEL, GOVERNANCE_TALK_CHANNEL
 are various channels used throughout the bot.
 Example usage is when posting information to Discord such as new events, new governance proposals, when members join, etc.
+
+YES_VOTE: The emoji used to vote yes on a proposal. Can be replaced with another emoji.
+
+NO_VOTE: The emoji used to vote no on a proposal.Can be replaced with another emoji.
+
+ABSTAIN_VOTE: The emoji used to abstain from voting on a proposal. Can be replaced with another emoji.
 
 new_proposal_emoji: The emoji used to inform the bot you wish to draft a new governance proposal.
 This is required when !v or !vote_draft command is invoked.
@@ -23,37 +33,39 @@ Each dictionary contains the following keys:
 
 MENU_COPY: A string containing the help menu for the bot. This menu lists all the available commands and their descriptions.
 
-CONFIG_ID_MAP: A dictionary containing the valid ids and keys that can be selected to be incremented or in rare cases decremented in the config.ini file.
-
-CONFIG_ABSOLUTE_PATH: The absolute path to the config.ini file.
-
 """
 
-CONFIG_ID_MAP: dict[str, str] = {"governance": "governance_id", "budget": "budget_id"}
-CONFIG_ABSOLUTE_PATH = "config/config.ini"
-
+# Preferred consts
 GENERAL_CHANNEL = "🌺│home"
 GOVERNANCE_CHANNEL = "improvement-props"
 GOVERNANCE_BUDGET_CHANNEL = "budgetary-props"
 GOVERNANCE_TALK_CHANNEL = "🐬│governance"
-new_proposal_emoji = "💡"
-CONTRIBUTORS_FILE_PATH = "./emotes/contributors.json"
-POSTED_EVENTS_FILE_PATH = "./events/posted_events.json"
+YES_VOTE = "👍"
+NO_VOTE = "👎"
+ABSTAIN_VOTE = "❌"
+
+# fallback consts
+FALLBACK_GENERAL_CHANNEL = "🐘│announcements"
+FALLBACK_GOVERNANCE_TALK_CHANNEL = "🌺│home"
+
+# There are no other forum channels within the server to use as a fallback
+FALLBACK_GOVERNANCE_BUDGET_CHANNEL = "improvement-props"
+FALLBACK_GOVERNANCE_CHANNEL = "budgetary-props"
+
+# Map the fallback consts to the preferred consts
+CONSTANT_FALLBACK_MAPPING = {
+    GENERAL_CHANNEL: FALLBACK_GENERAL_CHANNEL,
+    GOVERNANCE_CHANNEL: FALLBACK_GOVERNANCE_CHANNEL,
+    GOVERNANCE_BUDGET_CHANNEL: FALLBACK_GOVERNANCE_BUDGET_CHANNEL,
+    GOVERNANCE_TALK_CHANNEL: FALLBACK_GOVERNANCE_TALK_CHANNEL,
+}
 
 RULES_MESSAGE_ID = 1202059311681904661  # Set to ID of whatever message you want to be used as rules / to welcome a user
 DISCORD_ROLE_TRIGGERS = [
     {"name": "Client", "emoji_id": 1199583728129802322, "role": "Client Pod"},
-    {
-        "name": "Graphics & Design",
-        "emoji_id": 980752213347549234,
-        "role": "Graphics Pod",
-    },
+    {"name": "Graphics & Design", "emoji_id": 980752213347549234, "role": "Graphics Pod"},
     {"name": "Backend", "emoji_id": 846911453839228938, "role": "Backend Pod"},
-    {
-        "name": "Gameplay & Story Pod",
-        "emoji_id": 961338498525306980,
-        "role": "Gameplay & Story Pod",
-    },
+    {"name": "Gameplay & Story Pod", "emoji_id": 961338498525306980, "role": "Gameplay & Story Pod"},
     {"name": "Operations", "emoji_id": 945177584768004127, "role": "Policy & Ops Pod"},
     {"name": "Financial", "emoji_id": 887872297082449970, "role": "Donator"},
 ]
@@ -62,39 +74,38 @@ MENU_COPY = """
 __ 🌺 **Bloom Bot Commands** 🌺 __
 Commands can be prompted with **!**
 
-**!help**
+**/help**
 ```
 Procure this message. You knew that already, silly goose.
 ```
-**!add_contributor**
+**/add_contributor**
 ```
 Add a contributor to the list of contributors
 ```
-**!remove_contributor**
+**/remove_contributor**
 ```
 Remove a contributor from the list of contributors
 ```
-**!contributors**
+**/contributors**
 ```
 Display the list of contributors
 ```
-**!list_events**
+**/list_events**
 ```
 Display the list of events and their associated event ID
 ```
-**!remove_event**
+**/remove_event**
 ```
 Remove an event from the list of events
-Example: !remove_event 2583458345
+Example: /remove_event my_event
 
 ```
-**!v or !vote_draft**
+**/vote_draft**
 ```
 Draft a vote proposal
 ```
-**!publish_draft**
+**/publish_draft**
 ```
 Publish a draft that has been worked on.
-Example: !publish_draft Budget Proposal
 ```
 """