+++ emotes/commands.py
@@ -5,9 +5,9 @@ The function calls related to the commands are located in command_operations.py
 setup_contributor_commands is used so that all event commands can be loaded at once. instead of individually.
 """
 
+import discord
 from typing import List, Dict
 from discord.ext import commands
-from discord.ext.commands import Context
 from emotes.command_operations import (
     list_contributors,
     remove_contributor,
@@ -29,35 +29,41 @@ def setup_contrbitutor_commands(
     emoji_dicts (Dict[str, Dict[str, str]]): The emojis to use in the commands.
     """
 
-    @bot.command(name="contributors")
-    async def listcontributors(ctx: Context) -> None:
+    @bot.tree.command(name="contributors")
+    async def listcontributors(interaction: discord.Interaction) -> None:
         """
-        Invokes list_contributors which lists all the contributors in a specific guild/s.
+        Lists the contributos associated with this guild.
 
         Parameters:
-        ctx (Context): The context in which the command was called.
+        interaction (Interaction): The interaction of the command invocation.
         """
-        await list_contributors(ctx, contributors, emoji_dicts)
+        await list_contributors(interaction, contributors, emoji_dicts)
 
-    @bot.command(name="remove_contributor")
-    async def removecontributor(ctx: Context, user_mention: str) -> None:
+    @bot.tree.command(name="remove_contributor")
+    async def removecontributor(
+        interaction: discord.Interaction, user_mention: str
+    ) -> None:
         """
-        Invokes remove_contributor which removes a contributor if applicable contributor exists
-        and the user invoking the command has the authorization to do so.
+        Removes a contributor from the list of contributors.
 
         Parameters:
-        ctx (Context): The context in which the command was called.
+        interaction (Interaction): The interaction of the command invocation.
         user_mention (str): The mention of the user to remove.
         """
-        await remove_contributor(ctx, contributors, emoji_dicts, user_mention)
+        await remove_contributor(interaction, contributors, emoji_dicts, user_mention)
 
-    @bot.command(name="add_contributor")
-    async def addcontributor(ctx: Context) -> None:
+    @bot.tree.command(name="add_contributor")
+    async def addcontributor(
+        interaction: discord.Interaction, user_mention: str, emoji: str
+    ) -> None:
         """
-        Invokes add_contributor which adds a contributor,
-        if the user invoking the command has the authorization to do so.
+        Adds a contibutor to the list of contributors.
 
         Parameters:
-        ctx (Context): The context in which the command was called.
+        interaction (Interaction): The interaction of the command invocation.
+        user_mention (str): The mention of the user to be added as a contributor.
+        emoji (str): The emoji associated with the user.
         """
-        await add_contributor(ctx, contributors, emoji_dicts, bot)
+        await add_contributor(
+            interaction, user_mention, emoji, contributors, emoji_dicts
+        )