+++ events/event_operations.py
@@ -8,7 +8,8 @@ import json
 import requests
 import asyncio
 import os
-from consts.constants import POSTED_EVENTS_FILE_PATH, GENERAL_CHANNEL
+from consts.constants import GENERAL_CHANNEL
+from config.config import POSTED_EVENTS_FILE_PATH
 from shared.helpers import get_channel_by_name
 from datetime import datetime, timezone
 from typing import List, Optional, Any
@@ -73,7 +74,7 @@ def format_event(event: ScheduledEvent, guild_id: int) -> str:
         f"\n"
         f"To request someones attendance, react to this message with their emoji! \n"
         f"\n"
-        f":link:** Event Link** {event_url} :link:\n"
+        f":link: ** Event Link {event_url} **:link:\n"
     )
     return formatted_event