+++ gov/proposals.py
@@ -15,7 +15,13 @@ import config.config as cfg
 import discord
 from typing import Dict, Any, List, Tuple
 from discord.ext.commands import Bot
-from consts.constants import GOVERNANCE_BUDGET_CHANNEL, GOVERNANCE_CHANNEL
+from consts.constants import (
+    GOVERNANCE_BUDGET_CHANNEL,
+    GOVERNANCE_CHANNEL,
+    YES_VOTE,
+    NO_VOTE,
+    ABSTAIN_VOTE,
+)
 from consts.types import GOVERNANCE_ID_TYPE, BUDGET_ID_TYPE
 from logger.logger import logger
 
@@ -25,12 +31,15 @@ ongoing_votes: Dict[int, Dict[str, Any]] = {}
 
 
 # prepare the draft by setting the type, channel ID, and title based on the draft type
-async def prepare_draft(draft: Dict[str, Any]) -> Tuple[str, str, str]:
+async def prepare_draft(
+    guild: discord.Guild, draft: Dict[str, Any]
+) -> Tuple[str, str, str]:
     """
     Prepare the draft by setting the type, channel ID, and title based on the draft type.
     Increment the current ID and update the config file.
 
     Parameters:
+    guild (discord.Guild): The guild to search for the channel in.
     draft (Dict[str, Any]): The draft to be prepared.
 
     Returns:
@@ -47,7 +56,9 @@ async def prepare_draft(draft: Dict[str, Any]) -> Tuple[str, str, str]:
         cfg.update_id_values(
             cfg.current_budget_id, id_type
         )  # Update the governance ID in the config file
-        title = f"Bloom Budget Proposal (BBP) #{cfg.current_budget_id}: {draft['name']}"
+        title = (
+            f"Bloom Budget Proposal (BBP) #{cfg.current_budget_id}: {draft['title']}"
+        )
     else:
         id_type = GOVERNANCE_ID_TYPE
         channel_name = GOVERNANCE_CHANNEL
@@ -55,13 +66,15 @@ async def prepare_draft(draft: Dict[str, Any]) -> Tuple[str, str, str]:
         cfg.update_id_values(
             cfg.current_governance_id, id_type
         )  # Update the governance ID in the config file
-        title = f"Bloom Governance Proposal (BGP) #{cfg.current_governance_id}: {draft['name']}"
+        title = f"Bloom Governance Proposal (BGP) #{cfg.current_governance_id}: {draft['title']}"
 
     return id_type, channel_name, title
 
 
 # publish the draft by creating a thread with the prepared content and starting a vote timer
-async def publish_draft(draft: Dict[str, Any], bot: Bot, guild_id: int) -> None:
+async def publish_draft(
+    draft: Dict[str, Any], bot: Bot, guild_id: int, guild: discord.Guild
+):
     """
     Publish the draft by creating a thread with the prepared content and starting a vote timer.
 
@@ -70,8 +83,7 @@ async def publish_draft(draft: Dict[str, Any], bot: Bot, guild_id: int) -> None:
     bot (Bot): The bot instance.
     guild_id (int): The ID of the guild where the draft will be published.
     """
-    id_type, channel_name, title = await prepare_draft(draft)
-
+    id_type, channel_name, title = await prepare_draft(guild, draft)
     forum_channel = discord.utils.get(
         bot.get_guild(guild_id).channels, name=channel_name
     )
@@ -92,16 +104,15 @@ async def publish_draft(draft: Dict[str, Any], bot: Bot, guild_id: int) -> None:
     **__Background__**
     {draft["background"]}
 
-    **👍 Yes**
+    **{YES_VOTE} Yes**
 
-    **👎 Reassess**
+    **{NO_VOTE} Reassess**
 
-    **❌ Abstain**
+    **{ABSTAIN_VOTE} Abstain**
 
     Vote will conclude in 48h from now.
     """
     )
-
     thread_with_message = await forum_channel.create_thread(name=title, content=content)
 
     ongoing_votes[thread_with_message.thread.id] = {
@@ -110,10 +121,42 @@ async def publish_draft(draft: Dict[str, Any], bot: Bot, guild_id: int) -> None:
         "reassess_count": 0,
         "abstain_count": 0,
     }
+    await react_to_vote(thread_with_message.thread.id, bot, guild_id, channel_name)
 
     await vote_timer(
         thread_with_message.thread.id, bot, guild_id, channel_name, title, draft
     )
+
+
+async def react_to_vote(
+    thread_id: int, bot: Bot, guild_id: int, channel_name: str
+) -> None:
+    """
+    React to the published draft with the vote emojis.
+    This function is called when a draft is published.
+    This will ensure that when a draft is published, the vote emojis are added to the thread.
+
+    Parameters:
+    thread_id (int): The ID of the thread where the vote is taking place.
+    bot (Bot): The bot instance.
+    guild_id (int): The ID of the guild where the vote is taking place.
+    channel_name (str): The name of the channel where the vote is taking place.
+    """
+    guild = bot.get_guild(guild_id)
+    if not guild:
+        logger.error(f"Error: Guild with id {guild_id} not found.")
+        return
+
+    # Fetch the channel by name
+    channel = discord.utils.get(guild.channels, name=channel_name)
+    thread = channel.get_thread(thread_id)
+
+    # Fetch the initial message in the thread using the thread ID
+    message = await thread.fetch_message(thread_id)
+
+    await message.add_reaction(YES_VOTE)
+    await message.add_reaction(NO_VOTE)
+    await message.add_reaction(ABSTAIN_VOTE)
 
 
 async def vote_timer(
@@ -143,7 +186,7 @@ async def vote_timer(
     # Fetch the guild using the guild_id
     guild = bot.get_guild(guild_id)
     if not guild:
-        print(f"Error: Guild with id {guild_id} not found.")
+        logger.error(f"Error: Guild with id {guild_id} not found.")
         return
 
     # Fetch the channel by name
@@ -152,14 +195,13 @@ async def vote_timer(
 
     # Fetch the initial message in the thread using the thread ID
     message = await thread.fetch_message(thread_id)
-
     for reaction in message.reactions:
-        if str(reaction.emoji) == "👍":
-            ongoing_votes[message.id]["yes_count"] = reaction.count
-        elif str(reaction.emoji) == "👎":
-            ongoing_votes[message.id]["reassess_count"] = reaction.count
-        elif str(reaction.emoji) == "❌":
-            ongoing_votes[message.id]["abstain_count"] = reaction.count
+        if str(reaction.emoji) == f"{YES_VOTE}":
+            ongoing_votes[message.id]["yes_count"] = reaction.count - 1
+        elif str(reaction.emoji) == f"{NO_VOTE}":
+            ongoing_votes[message.id]["reassess_count"] = reaction.count - 1
+        elif str(reaction.emoji) == f"{ABSTAIN_VOTE}":
+            ongoing_votes[message.id]["abstain_count"] = reaction.count - 1
 
     # Check the result and post it
     result_message = f"Vote for '{title}' has concluded:\n\n"