+++ main.py
@@ -24,7 +24,7 @@ from emotes.commands import setup_contrbitutor_commands
 from events.commands import setup_event_commands
 from events.events import setup_event_events
 from events.event_operations import load_posted_events
-from consts.constants import CONTRIBUTORS_FILE_PATH, new_proposal_emoji
+from config.config import CONTRIBUTORS_FILE_PATH
 from shared.events import setup_shared_events
 
 
@@ -34,15 +34,13 @@ class Bot:
         with open(CONTRIBUTORS_FILE_PATH, "r") as json_file:
             self.data = json.load(json_file)
             self.contributors = {
-                "Bloom Studio": self.data["servers"]["Bloom Studio"]["contributors"],
+                "priv-server": self.data["servers"]["priv-server"]["contributors"],
                 "Bloom Collective": self.data["servers"]["Bloom Collective"][
                     "contributors"
                 ],
             }
             self.emoji_dicts = {
-                "Bloom Studio": self.data["servers"]["Bloom Studio"][
-                    "emoji_dictionary"
-                ],
+                "priv-server": self.data["servers"]["priv-server"]["emoji_dictionary"],
                 "Bloom Collective": self.data["servers"]["Bloom Collective"][
                     "emoji_dictionary"
                 ],
@@ -53,12 +51,12 @@ class Bot:
         intents.message_content = True
         intents.reactions = True
         intents.members = True
-        self.bot = commands.Bot(command_prefix="!", intents=intents)
+        self.bot = commands.Bot(command_prefix="", intents=intents)
 
         # Setup commands and events for the bot
         setup_gov_commands(self.bot)
         setup_contrbitutor_commands(self.bot, self.contributors, self.emoji_dicts)
-        setup_shared_events(self.bot, self.data, proposals, new_proposal_emoji)
+        setup_shared_events(self.bot, self.data, proposals)
         self.bot.posted_events = load_posted_events()
         setup_event_commands(self.bot)
         setup_event_events(self.bot)